[Minuit2] Code refactoring to reduce interface and reuse helper class
